requirements.txt
tensorflow
keras
numpy
tqdm
Pillow==9.5.0
opencv-python
scikit-learn